move PathFactory to io module
